refactor: auth
- merge env files
- moved some files around until it felt good
