+++ .env.example
@@ -1,15 +1,15 @@
-# required
-NEXT_PUBLIC_REPO_LINK=https://github.com/ljosberinn/personal-react-boilerplate/tree/next
-NEXT_PUBLIC_PROJECT_NAME=Brand Name
+# i18n
 NEXT_PUBLIC_ENABLED_LANGUAGES=en,de
 
 # sentry
 NEXT_PUBLIC_SENTRY_DSN=
 
 # auth
-NEXT_PUBLIC_ENABLED_PROVIDER=github,google,facebook,discord,local
+NEXT_PUBLIC_ENABLED_PROVIDER=local,github,google,facebook,discord
 
-TOKEN_SECRET=this-is-a-secret-value-with-at-least-32-characters
+SESSION_COOKIE_SECRET=this-is-a-secret-value-with-at-least-32-characters
+# defaults to 8 hours; value must be in SECONDS
+SESSION_LIFETIME=
 
 GITHUB_CLIENT_ID=
 GITHUB_CLIENT_SECRET=
@@ -22,6 +22,3 @@ FACEBOOK_CLIENT_SECRET=
 
 DISCORD_CLIENT_ID=
 DISCORD_CLIENT_SECRET=
-
-# other
-NEXT_PUBLIC_DISCORD_LINK=
