+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "personal-react-boilerplate",
+  "name": "next-with-batteries",
   "version": "1.0.0",
   "scripts": {
     "dev": "next",
