+++ pages/api/v1/auth/logout.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { SESSION_COOKIE_NAME } from '../../../../src/server/auth/authConstants';
+import { SESSION_COOKIE_NAME } from '../../../../src/constants';
 import { removeCookie } from '../../../../src/server/auth/cookie';
 import { OK } from '../../../../src/utils/statusCodes';
