+++ pages/api/v1/user/me.ts
@@ -1,14 +1,11 @@
 import nextConnect from 'next-connect';
 
-import {
-  authNSecurityMiddleware,
-  expectJSONBodyMiddleware,
-} from '../../../../src/server/auth/middlewares';
+import { SESSION_COOKIE_NAME } from '../../../../src/constants';
+import { authNSecurityMiddleware } from '../../../../src/server/auth/middlewares';
 import { AuthenticatedRequest } from '../../../../src/server/auth/types';
 
 export default nextConnect()
   .use(authNSecurityMiddleware)
-  .use(expectJSONBodyMiddleware)
   .get<AuthenticatedRequest>((req, res) => {
-    res.json(req.session);
+    res.json(req[SESSION_COOKIE_NAME]);
   });
