+++ pages/api/v1/user/todo.ts
@@ -1,13 +1,9 @@
 import nextConnect from 'next-connect';
 
-import {
-  authNSecurityMiddleware,
-  expectJSONBodyMiddleware,
-} from '../../../../src/server/auth/middlewares';
+import { authNSecurityMiddleware } from '../../../../src/server/auth/middlewares';
 
 export default nextConnect()
   .use(authNSecurityMiddleware)
-  .use(expectJSONBodyMiddleware)
   .get(async (_, res) => {
     try {
       const response = await fetch(
