--- src/server/auth/authConstants.ts
@@ -1,3 +1,0 @@
-export const SESSION_COOKIE_NAME = 'session';
-
-export const MAX_AGE = 60 * 60 * 8 * 1000; // 8 hours
