+++ src/server/auth/types.ts
@@ -1,7 +1,7 @@
 import { Profile } from 'passport';
 
 import { User } from '../../client/context/AuthContext/AuthContext';
-import { SESSION_COOKIE_NAME } from './authConstants';
+import { SESSION_COOKIE_NAME } from '../../constants';
 
 export interface AuthenticatedRequest {
   [SESSION_COOKIE_NAME]: User;
